+++ src/frontends/lean/info_manager.cpp
@@ -77,10 +77,6 @@ hole_info_data const & to_hole_info_data(info_data const & d) {
 vm_obj_format_info const * is_vm_obj_format_info(info_data const & d) {
     return dynamic_cast<vm_obj_format_info const *>(d.raw());
 }
-widget_info const * is_widget_info(info_data const & d) {
-    return dynamic_cast<widget_info const *>(d.raw());
-}
-
 
 class term_goal_data : public info_data_cell {
     pos_info m_pos;
@@ -121,89 +117,12 @@ void vm_obj_format_info::report(io_state_stream const & ios, json & record) cons
 }
 #endif
 
-json widget_info::to_json() const {
-    vdom * vd = const_cast<vdom*>(&m_vdom);
-    return vd->to_json();
-}
-
-void widget_info::report(io_state_stream const & ios, json & record) const {
-    if (!get_global_module_mgr()->get_report_widgets()) { return; }
-    mutex * mp = const_cast<mutex *>(&m_mutex);
-    lock_guard<mutex> _(*mp);
-    vm_state S(m_env, ios.get_options());
-    scope_vm_state scope(S);
-    pending_tasks pts;
-    set_pending_tasks(&pts);
-    record["widget"]["html"] = to_json();
-    unset_pending_tasks();
-    any(pts, [] (list<unsigned> route) => {
-        // perform the update. notify the server somehow, probably through the log tree.
-    })
-    // [todo] for each task in pts notify the logtree that the state changed whenever one of the tasks completes.
-
-    // sketch: I would rather this does something like 'task.first_to_complete(gs)`
-    for (auto pt:pts) {
-        gtask g = pt.first;
-        auto route = pt.second.first;
-        auto handler_id = pt.second.second;
-        taskq().submit(task_builder([g, route, handler_id] {
-            auto r = g.get();
-            auto res = c->handle_event(route, handler_id, mk_vm_unit());
-            lean_assert(!res);
-            return unit{};
-        }).depends_on(g).build());
-    }
-}
-
-void widget_info::update(io_state_stream const & ios, json const & message, json & record) {
-    if (!get_global_module_mgr()->get_report_widgets()) { return; }
-    lock_guard<mutex> _(m_mutex);
-    vm_state S(m_env, ios.get_options());
-    scope_vm_state scope(S);
-    unsigned handler_idx = message["handler"]["h"];
-    json j_route = message["handler"]["r"]; // an array with the root index at the _back_.
-    list<unsigned> route; // now root index is at the _front_.
-    for (json::iterator it = j_route.begin(); it != j_route.end(); ++it) {
-      route = cons(unsigned(*it), route);
-    }
-    route = tail(route); // disregard the top component id because that is the root component
-    json j_args = message["args"];
-    component_instance * c = const_cast<component_instance *>(dynamic_cast<component_instance *>(m_vdom.raw()));
-    vm_obj vm_args;
-    std::string arg_type = j_args["type"];
-    if (arg_type == "unit") {
-        vm_args = mk_vm_unit();
-    } else if (arg_type == "string") {
-          std::string arg = j_args["value"];
-          vm_args = to_obj(arg);
-    } else {
-        throw exception("expecting arg_type to be either 'unit' or 'string' but was '" + arg_type + "'");
-    }
-    try {
-        optional<vm_obj> result = c->handle_event(route, handler_idx, vm_args);
-        record["widget"]["html"] = to_json();
-        if (result) {
-            record["status"] = "edit";
-            record["action"] = to_string(*result);
-        } else {
-            record["status"] = "success";
-        }
-    } catch (const invalid_handler & e) {
-        record["status"] = "invalid_handler";
-    }
-}
-
 info_data mk_type_info(expr const & e) { return info_data(new type_info_data(e)); }
 
 info_data mk_identifier_info(name const & full_id) { return info_data(new identifier_info_data(full_id)); }
 
 info_data mk_vm_obj_format_info(environment const & env, vm_obj const & thunk) {
     return info_data(new vm_obj_format_info(env, thunk));
-}
-
-info_data mk_widget_info(environment const & env, vm_obj const & props, vm_obj const & widget) {
-    vdom c = vdom(new component_instance(widget, props));
-    return info_data(new widget_info(env, c));
 }
 
 info_data mk_hole_info(tactic_state const & s, expr const & hole_args, pos_info const & begin, pos_info end) {
@@ -295,14 +214,6 @@ void info_manager::add_term_goal(pos_info const & pos, tactic_state const & s) {
     add_info(pos, mk_term_goal(pos, s));
 }
 
-void info_manager::add_widget_info(pos_info pos, vm_obj const & props, vm_obj const & widget) {
-#ifdef LEAN_NO_INFO
-    return;
-#endif
-    add_info(pos, mk_widget_info(tactic::to_state(props).env(), props, widget));
-}
-
-
 #ifdef LEAN_JSON
 void info_manager::get_info_record(environment const & env, options const & o, io_state const & ios, pos_info pos,
                                    json & record, std::function<bool (info_data const &)> pred) const {
@@ -322,26 +233,6 @@ optional<list<info_data>> info_manager::get_info(pos_info pos) const{
     optional<list<info_data>> result;
     if (ds) { result = some(*ds); }
     return result;
-}
-
-bool info_manager::update_widget(environment const & env, options const & o, io_state const & ios, pos_info pos, json & record, json const & message) const {
-    type_context_old tc(env, o);
-    io_state_stream out = regular(env, ios, tc).update_options(o);
-    auto ds = get_info(pos);
-    if (!ds) {
-        record["status"] = "error";
-        record["message"] = "could not find a widget at the given position";
-        return false;
-     }
-    for (auto & d : *ds) {
-        const widget_info* cw = is_widget_info(d);
-        widget_info * w = const_cast<widget_info*>(cw);
-        if (w) {
-            w->update(out, message, record);
-            return true;
-        }
-    }
-    return false;
 }
 
 LEAN_THREAD_PTR(info_manager, g_info_m);
